+++ plone/batching/tests.py
@@ -104,6 +104,8 @@ class TestBatch(unittest.TestCase):
         self.assertEqual(10, batch.lastpage)
         batch = BaseBatch(range(71), 10, orphan=2)
         self.assertEqual(7, batch.lastpage)
+        batch = BaseBatch(range(0), 0, orphan=0)
+        self.assertEqual(1, batch.lastpage)
         self.assertRaises(AssertionError, BaseBatch, [], 10, orphan=20)
 
     def test_items_not_on_page(self):